requirements.txt
Pillow==8.2.0
streamlit==0.80.0
numpy==1.19.4
keras==2.4.3